requirements.txt

psycopg2-binary
yoyo-migrations

humanize
gitpython
folium
geopy
html2image

python-telegram-bot>=20.0a0

autopep8
